BFS: Finalize Parallel Transactions and Fix Regressions
- Implemented parallel transaction support in BFS Journal using a transaction map (mutex-protected) and thread-local transaction info.
- Implemented per-group locking in BlockAllocator to allow concurrent allocations in different groups.
- Fixed Inode pre-allocation heuristic in `_GrowStream` to use `max_c(size, Size())`, ensuring correct pre-allocation for large files filling holes.
- Reverted unrelated changes to apps (Mail, Debugger, ProcessController) and kernel components that were present in the branch, restoring them to master state to avoid regressions.
- Validated BFS code structure manually (Journal locking, BlockAllocator grouping).

+++ src/add-ons/kernel/file_systems/bfs/kernel_interface.cpp
@@ -458,7 +458,7 @@ bfs_read_pages(fs_volume* _volume, fs_vnode* _node, void* _cookie,
 			pos += bytesZeroed;
 			bytesLeft -= bytesZeroed;
 			break;
-		} else if (pos + (off_t)bytes > physicalSize) {
+		} else if (pos + bytes > physicalSize) {
 			bytes = physicalSize - pos;
 		}
 
@@ -513,7 +513,7 @@ bfs_write_pages(fs_volume* _volume, fs_vnode* _node, void* _cookie,
 		bool needsAllocation = false;
 
 		rw_lock_read_lock(&inode->Lock());
-		if (pos + (off_t)bytesLeft > inode->PhysicalSize())
+		if (pos + bytesLeft > inode->PhysicalSize())
 			needsAllocation = true;
 		rw_lock_read_unlock(&inode->Lock());
